+++ explorer/src/lib/chain-info/index.js
@@ -1,3 +1,4 @@
+export { default as addCountAndUnique } from "./addCountAndUnique";
 export { default as calculateStats } from "./calculateStats";
 export { default as transformBlock } from "./transformBlock";
 export { default as transformSearchResult } from "./transformSearchResult";