+++ explorer/src/lib/node-location/node-location.d.ts
@@ -1,0 +1,8 @@
+type NodeLocation = {
+  lat: number;
+  lon: number;
+  city: string;
+  count: number;
+  country: string;
+  countryCode: string;
+};